Very significant overhaul to make it work robustly when released
There were some problems which could always cause silent problems, and
other problems which caused problems when the software is released.

* Global variables are initialized during "importing".  I removed most
  of these, though some still rename.  It seems like an attempt to
  implement singletons, but it is very fragile and depending on
  various python namespace and scope rules it can backfire and lose
  values of global variables.

* Importing * caused problems.  I removed many cases of that.

* In a few places exceptions are used for flow control.  This is a
  real problem because it is so hard to get right, and in fact it
  caused several silent failures, and other places where it was very
  hard to debug what was happening with real hard failures, since the
  except: blocks were not loud enough.  I put stronger diagnostics in
  some of the except: blocks but have not yet moved away from that
  overall practice.

Finally: the reorganization to better use the python "hatch" approach
to building is a nice feature since earlier versions of nsiqcppstyle,
but it was not yet complete.

I did some rearranging of the sources to match recommended
pyproject.toml approaches.  This includes specifying the executable
script in pyproject.toml as:

[project.scripts]
nsiqcppstyle = "nsiqcppstyle.nsiqcppstyle_exe:main"

If anyone uses this on windows they will need to follow the
pyproject.toml approach to make a GUI script.  I am not doing that
because I have no way of testing it for Windows.

I also tried to make things work well with the Fedora packaging
guidelines.  This is hard to do because we are in the midst of a
medium-scope switch in Fedora guidelines, and in this RPM spec file I
am using the (slightly) older guidelines from 2021 (instead of 2023)
so that we can function on el8 and el9.